requirements.txt

flask
pillow
requests
beautifulsoup4
